# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Zorkin Store API",
        "contact": {}
    },
    "paths": {
...
        "/api/admin/presets": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new preset with its items",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Preset"
                ],
                "summary": "Create a new preset",
                "parameters": [
                    {
                        "description": "Preset data",
                        "name": "preset",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Neimess_zorkin-store-project_internal_transport_http_restHTTP_preset_dto.PresetRequest"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Neimess_zorkin-store-project_pkg_httputils.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_Neimess_zorkin-store-project_pkg_httputils.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_Neimess_zorkin-store-project_pkg_httputils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Neimess_zorkin-store-project_pkg_httputils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/admin/presets/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Preset"
                ],
                "summary": "Delete preset",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Preset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Neimess_zorkin-store-project_pkg_httputils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Neimess_zorkin-store-project_pkg_httputils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Neimess_zorkin-store-project_pkg_httputils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/presets": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Preset"
                ],
                "summary": "List presets basic info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_Neimess_zorkin-store-project_internal_transport_http_restHTTP_preset_dto.PresetShortResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Neimess_zorkin-store-project_pkg_httputils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/presets/detailed": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Preset"
                ],
                "summary": "List presets with items",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_Neimess_zorkin-store-project_internal_transport_http_restHTTP_preset_dto.PresetResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Neimess_zorkin-store-project_pkg_httputils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/presets/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Preset"
                ],
                "summary": "Get preset by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Preset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Neimess_zorkin-store-project_internal_transport_http_restHTTP_preset_dto.PresetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Neimess_zorkin-store-project_pkg_httputils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Neimess_zorkin-store-project_pkg_httputils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Neimess_zorkin-store-project_pkg_httputils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_Neimess_zorkin-store-project_internal_transport_http_restHTTP_preset_dto.PresetRequest": {
            "type": "object",
            "required": [
                "items",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Neimess_zorkin-store-project_internal_transport_http_restHTTP_preset_dto.PresetRequestItem"
                    }
                },
                "name": {
                    "type": "string",
                    "minLength": 2
                },
                "total_price": {
                    "type": "number"
                }
            }
        },
        "github_com_Neimess_zorkin-store-project_internal_transport_http_restHTTP_preset_dto.PresetRequestItem": {
            "type": "object",
            "required": [
                "product_id"
            ],
            "properties": {
                "product_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_Neimess_zorkin-store-project_internal_transport_http_restHTTP_preset_dto.PresetResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-06-20T15:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Полный комплект для ванной комнаты"
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/image.png"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Neimess_zorkin-store-project_internal_transport_http_restHTTP_preset_dto.PresetResponseItem"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "Комплект для ванной"
                },
                "preset_id": {
                    "type": "integer",
                    "example": 1
                },
                "total_price": {
                    "type": "number",
                    "example": 15000
                }
            }
        },
        "github_com_Neimess_zorkin-store-project_internal_transport_http_restHTTP_preset_dto.PresetResponseItem": {
            "type": "object",
            "properties": {
                "product": {
                    "$ref": "#/definitions/github_com_Neimess_zorkin-store-project_internal_transport_http_restHTTP_preset_dto.ProductSummary"
                }
            }
        },
        "github_com_Neimess_zorkin-store-project_internal_transport_http_restHTTP_preset_dto.PresetShortResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-06-20T15:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Для ванной комнаты"
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/image.png"
                },
                "name": {
                    "type": "string",
                    "example": "Комплект для ванной"
                },
                "preset_id": {
                    "type": "integer",
                    "example": 1
                },
                "total_price": {
                    "description": "может быть опциональным",
                    "type": "number",
                    "example": 15000
                }
            }
        },
        "github_com_Neimess_zorkin-store-project_internal_transport_http_restHTTP_preset_dto.ProductSummary": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 10
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/shampoo.png"
                },
                "name": {
                    "type": "string",
                    "example": "Шампунь"
                },
                "price": {
                    "type": "number",
                    "example": 499
                }
            }
        },
...
        "github_com_Neimess_zorkin-store-project_pkg_httputils.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type **\"Bearer \u003cJWT\u003e\"** here",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}